+++ src/components/description/Description.js
@@ -1,11 +1,30 @@
 import React from 'react';
 
 export default function Description(props) {
-  const { description } = props;
+  const { description, metadata } = props;
 
   return (
-    <div>
-      <p className="details">{description}</p>
+    <div className="description">
+      <p>{description}</p>
+
+      {Object.entries(metadata).map(([layerName, layerMetadata]) => (
+        <details key={layerName}>
+          <summary>{layerName}</summary>
+          <div className="metadata-container">
+            <table>
+              <tbody>
+                {Object.entries(layerMetadata)
+                  .map(([key, value]) => (
+                    <tr key={key}>
+                      <th title={key}>{key}</th>
+                      <td title={value}>{value}</td>
+                    </tr>
+                  ))}
+              </tbody>
+            </table>
+          </div>
+        </details>
+      ))}
     </div>
   );
 }