+++ src/components/description/DescriptionSubscriber.js
@@ -1,6 +1,6 @@
-import React, { useCallback, useEffect } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import PubSub from 'pubsub-js';
-import { RASTER_ADD } from '../../events';
+import { METADATA_ADD } from '../../events';
 import TitleInfo from '../TitleInfo';
 import Description from './Description';
 
@@ -8,16 +8,20 @@ export default function DescriptionSubscriber(props) {
   const { description, onReady, removeGridComponent } = props;
 
   const onReadyCallback = useCallback(onReady, []);
+  const [metadata, setMetadata] = useState({});
+
+  const handleMetadataAdd = useCallback((msg, { name: layerName, metadata: layerMetadata }) => {
+    setMetadata({
+      ...metadata,
+      [layerName]: layerMetadata,
+    });
+  }, [metadata, setMetadata]);
 
   useEffect(() => {
-    function handleRasterAdd(msg, raster) {
-      // eslint-disable-next-line
-          console.log(raster);
-    }
+    const metadataAddToken = PubSub.subscribe(METADATA_ADD, handleMetadataAdd);
     onReadyCallback();
-    const token = PubSub.subscribe(RASTER_ADD, handleRasterAdd);
-    return () => PubSub.unsubscribe(token);
-  }, [onReadyCallback]);
+    return () => PubSub.unsubscribe(metadataAddToken);
+  }, [onReadyCallback, handleMetadataAdd]);
 
   return (
     <TitleInfo
@@ -27,6 +31,7 @@ export default function DescriptionSubscriber(props) {
     >
       <Description
         description={description}
+        metadata={metadata}
       />
     </TitleInfo>
   );