+++ src/components/layer-controller/LayerController.js
@@ -16,10 +16,9 @@ import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import ChannelController from './ChannelController';
 import LayerOptions from './LayerOptions';
 
-import { LAYER_ADD, LAYER_CHANGE } from '../../events';
+import { LAYER_ADD, LAYER_CHANGE, METADATA_ADD } from '../../events';
 import reducer from './reducer';
 import { useExpansionPanelStyles } from './styles';
-
 
 async function initLoader(imageData) {
   const { type, url, metadata } = imageData;
@@ -74,6 +73,12 @@ export default function LayerController({ imageData, layerId, handleLayerRemove 
         loader,
         layerProps: DEFAULT_LAYER_PROPS,
       });
+      if (loader.getMetadata) {
+        PubSub.publish(METADATA_ADD, {
+          name: imageData.name,
+          metadata: loader.getMetadata(),
+        });
+      }
       // Add channel on image add automatically
       dispatch({
         type: 'ADD_CHANNEL',